# swagger.yaml
openapi: 3.0.0

info:
  title: Textile

components:
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT    

# security:
#   - bearerAuth: []
# summary: Submit email address and password to login
servers:
  - url: http://localhost:3001/
    description: Local server

paths:               
  /register:
    post:      
      tags:
       - "AUTHENDICATION"
      description: ""
      operationId: register      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile:
                  type: string
                  example: "1234567890"
                email:
                  type: string
                  example: "mhariharanms@gmail.com"
                password:
                  type: string
                  example: "Pass@123"
                name:
                  type: string
                  example: "Jack"
                userRole: 
                  type: string
                  example: "SELLER"           
              required:   
                - mobile            
                - password
                - email
                - name
                - userRole
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
  
  /login:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: login      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string  
                  example: "1234567890"                 
                password:
                  type: string    
                  example: "Pass@123"                           
              required:
                - username
                - password                
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
    
  /logout:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: logout      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string                                      
              required:
                - token               
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
 
  /forgotPassword:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: forgotpassword      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string                                   
              required:
                - username                              
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
  
  /changePassword:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: changePassword      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string                                  
              required:
                - username  
                - password
                - confirmPassword                            
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error

  /resendOtp:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: resendOtp      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile:
                  type: string                                                          
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
  
  /refreshToken:
    post:
      tags:
      - "AUTHENDICATION"
      description: ""
      operationId: refreshToken      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                token:
                  type: string 
              required:
                - userId
                - token                                            
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
   
  /deleteUser:
    delete:
      security:
        - bearerAuth: []
      tags:
      - "ADMIN"
      description: ""
      operationId: deleteUser      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string                          
              required:
                - userId                                     
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error
 
  /getAllUsers:
    post:
      security:
        - bearerAuth: []
      tags:
      - "ADMIN"
      description: ""
      operationId: getAllUsers 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filterObj:
                  type: object
                  properties:
                    userRole: 
                      type: string
                    searchValue: 
                      type: string
                    isActive:
                      type: boolean
                    userStatus:
                      type: string
                sortObj:
                  type: object
                  properties:
                    createdAt: 
                      type: number   
                page:
                  type: number
                limit:
                  type: number                        
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error  
  
  /getAllProducts:
    post:
      security:
        - bearerAuth: []
      tags:
      - "ADMIN"
      description: ""
      operationId: getAllProducts 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filterObj:
                  type: object
                  properties:
                    searchValue: 
                      type: string
                sortObj:
                  type: object
                  properties:
                    createdAt: 
                      type: number   
                page:
                  type: number
                limit:
                  type: number                        
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error  
  
  /getAllSales:
    post:
      security:
        - bearerAuth: []
      tags:
      - "ADMIN"
      description: ""
      operationId: getAllSales 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filterObj:
                  type: object
                  properties:
                    searchValue: 
                      type: string
                sortObj:
                  type: object
                  properties:
                    createdAt: 
                      type: number   
                page:
                  type: number
                limit:
                  type: number                        
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request Error
        "401":
          description: Unauthorized Error  
  
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid  